Rename Def files and update their import
