+++ qcodes/instrument_drivers/Keysight/KtMAwg.py
@@ -1,4 +1,4 @@
-from .KtMAwg_Defs import *
+from .KtMAwgDefs import *
 
 import ctypes
 from functools import partial
